test(telegram): co-locate transport tests in telegram-api.test.ts
3 of the 4 extracted subsystems have co-located test files (turn-queue,
context-buffer, preview); telegram-api's tests stayed inside the 1799-line
telegram.test.ts. Move the chunkText / sendMessage-fallback / callApi-pipeline
/ file-download describes into test/telegram-api.test.ts so the test layout
mirrors the module layout. Pure move; 77 tests unchanged.
